requirements.txt
Flask>=2.2.3
Werkzeug>=3.1.3
PyYaml>=6.0.2
Jinja2>=3.1.6
flask_login>=0.6.3
bcrypt>=4.3.0
Flask_bcrypt>=1.0.1
dynaconf>=3.2.10
DebugToolbarExtension>=0.16.0
flask_sqlalchemy>=3.1.1
flask_wtf>=1.2.2
email_validator>=2.2.0
wtforms>=3.2.1